Database works with login
The problem was the following: 1. the ; in the ini files comments out the line. it needed to be uncommented. 2. The file php.ini needed to exist, development and production on the end of them made them not be read by php.

Additionally any file that uses php code must use .php extension. Putting php code into a .html will not be processed.