SQL Injection test made for login
I made this to make presenting login SQL injection easier to understand and to make testing different SQL Injection statements easier as well. It is just a copy of Index where stuff on the page is updated via a selection box. To make more examples, just make a new option for them and then add some javascript to update the page appropriately.