Added another injection test
query does not seem to handle batch statements. I tried both DROP TABLE User and DROP TABLE User; The password check seems to work if i use OR 1=1;, but it won't run the DROP. This means it only works on exec, which means the only place it will work is in the forgot password use case. We could also arbitrarily add an exec in before doing the query for login if we choose to do so.