Moved css and svg to resources and cleaned up some code
Standardized the code a bit more. Added a few more comments here and there despite it not being my main goal. Moved css and svg to resources and changed all the header statements to reflect this. Fixed an error in the logout button for edit_account. Removed some unused files from resources folder.