Wrapped INSERT statements in a transaction for a notable speed increase
I found the page was not loading in time before timing out, so I looked into the issue. Apparently each insert was being logged, but wrapping it in a transaction results in 1 log for everything in the transaction instead of around 100 or so. It would only get slower the more that was added without this.