Added navigation back to main site and some very basic navigation to DBcontents
I tried to align the button to the right of the header of each section in DBcontents, but it looked a bit off. Another downside with my implementation is that the history gets filled each time you navigate to a section on the page. Javascript might be a better plan to handle this than pure html.